# [full-ci] enhancement: add more kql spec tests and simplify ast normalization