--- a/ocis-pkg/tools/tools.go
+++ b/ocis-pkg/tools/tools.go
@@ -1,7 +1,8 @@
-// +build tools
+//go:build tools
 
 package tools
 
 import (
+	_ "github.com/mna/pigeon"
 	_ "github.com/onsi/ginkgo/ginkgo"
 )