--- a/services/search/pkg/query/kql/cast.go
+++ b/services/search/pkg/query/kql/cast.go
@@ -23,8 +23,24 @@ func toNodes[T ast.Node](in interface{}) ([]T, error) {
 	switch v := in.(type) {
 	case []T:
 		return v, nil
+	case T:
+		return []T{v}, nil
+	case []interface{}:
+		var ts []T
+		for _, inter := range v {
+			n, err := toNodes[T](inter)
+			if err != nil {
+				return nil, err
+			}
+
+			ts = append(ts, n...)
+		}
+		return ts, nil
+	case nil:
+		return nil, nil
 	default:
-		return nil, fmt.Errorf("can't convert '%T' to []ast.Node", in)
+		var t T
+		return nil, fmt.Errorf("can't convert '%T' to '%T'", in, t)
 	}
 }