--- a/services/search/pkg/query/kql/const.go
+++ b/services/search/pkg/query/kql/const.go
@@ -1,8 +1,1 @@
 package kql
-
-// The operator node value definition
-const (
-	BoolAND = "AND"
-	BoolOR  = "OR"
-	BoolNOT = "NOT"
-)