--- a/services/search/pkg/query/kql/const.go
+++ b/services/search/pkg/query/kql/const.go
@@ -1,1 +1,0 @@
-package kql