--- a/services/search/pkg/query/kql/error.go
+++ b/services/search/pkg/query/kql/error.go
@@ -1,10 +1,30 @@
 package kql
 
+import (
+	"fmt"
+
+	"github.com/owncloud/ocis/v2/services/search/pkg/query/ast"
+)
+
 // StartsWithBinaryOperatorError records an error and the operation that caused it.
 type StartsWithBinaryOperatorError struct {
-	Op string
+	Node *ast.OperatorNode
 }
 
-func (e *StartsWithBinaryOperatorError) Error() string {
-	return "the expression can't begin from a binary operator: '" + e.Op + "'"
+func (e StartsWithBinaryOperatorError) Error() string {
+	return "the expression can't begin from a binary operator: '" + e.Node.Value + "'"
+}
+
+// NamedGroupInvalidNodesError records an error and the operation that caused it.
+type NamedGroupInvalidNodesError struct {
+	Node ast.Node
+}
+
+func (e NamedGroupInvalidNodesError) Error() string {
+	return fmt.Errorf(
+		"'%T' - '%v' - '%v' is not valid",
+		e.Node,
+		ast.NodeKey(e.Node),
+		ast.NodeValue(e.Node),
+	).Error()
 }