--- a/services/search/pkg/query/kql/factory.go
+++ b/services/search/pkg/query/kql/factory.go
@@ -40,33 +40,8 @@ func buildAST(n interface{}, text []byte, pos position) (*ast.Ast, error) {
 
 	return &ast.Ast{
 		Base:  b,
-		Nodes: nodes,
+		Nodes: connectNodes(DefaultConnector{sameKeyOPValue: BoolOR}, nodes...),
 	}, nil
-}
-
-func buildNodes(head, tail interface{}) ([]ast.Node, error) {
-	headNode, err := toNode[ast.Node](head)
-	if err != nil {
-		return nil, err
-	}
-
-	if tail == nil {
-		return []ast.Node{headNode}, nil
-	}
-
-	tailNodes, err := toNodes[ast.Node](tail)
-	if err != nil {
-		return nil, err
-	}
-
-	allNodes := []ast.Node{headNode}
-
-	connectionNode := incorporateNode(headNode, tailNodes...)
-	if connectionNode != nil {
-		allNodes = append(allNodes, connectionNode)
-	}
-
-	return append(allNodes, tailNodes...), nil
 }
 
 func buildStringNode(k, v interface{}, text []byte, pos position) (*ast.StringNode, error) {
@@ -184,6 +159,6 @@ func buildGroupNode(k, n interface{}, text []byte, pos position) (*ast.GroupNode
 	return &ast.GroupNode{
 		Base:  b,
 		Key:   key,
-		Nodes: nodes,
+		Nodes: connectNodes(DefaultConnector{sameKeyOPValue: BoolAND}, nodes...),
 	}, nil
 }