--- a/services/search/pkg/query/kql/gen.go
+++ b/services/search/pkg/query/kql/gen.go
@@ -1,3 +1,3 @@
 package kql
 
-//go:generate go run github.com/mna/pigeon -o dictionary_gen.go dictionary.peg
+//go:generate go run github.com/mna/pigeon -optimize-grammar -optimize-parser -o dictionary_gen.go dictionary.peg