… changelog/unreleased/kql-fix-unary.md
… changelog/unreleased/kql-search-query-language.md
… go.sum
… services/search/pkg/query/kql/dictionary_gen.go
… services/search/pkg/query/kql/kql_test.go
… services/search/pkg/query/kql/normalize.go
… services/search/pkg/query/kql/normalize_test.go
… vendor/github.com/araddon/dateparse/.travis.yml
… vendor/github.com/araddon/dateparse/LICENSE
… vendor/github.com/araddon/dateparse/README.md
… vendor/github.com/araddon/dateparse/parseany.go
… vendor/github.com/mna/pigeon/.editorconfig
… vendor/github.com/mna/pigeon/.gitattributes
… vendor/github.com/mna/pigeon/.gitignore
… vendor/github.com/mna/pigeon/.travis.yml
… vendor/github.com/mna/pigeon/CONTRIBUTING.md
… vendor/github.com/mna/pigeon/LICENSE
… vendor/github.com/mna/pigeon/Makefile
… vendor/github.com/mna/pigeon/README.md
… vendor/github.com/mna/pigeon/TODO
… vendor/github.com/mna/pigeon/ast/ast.go
… vendor/github.com/mna/pigeon/ast/ast_optimize.go
… vendor/github.com/mna/pigeon/ast/ast_walk.go
… vendor/github.com/mna/pigeon/builder/builder.go
… vendor/github.com/mna/pigeon/builder/generated_static_code.go
… vendor/github.com/mna/pigeon/builder/generated_static_code_range_table.go
… vendor/github.com/mna/pigeon/builder/static_code.go
… vendor/github.com/mna/pigeon/builder/static_code_range_table.go
… vendor/github.com/mna/pigeon/doc.go
… vendor/github.com/mna/pigeon/main.go