## Description
follow up pr for the ocis kql integration, @butonic found a ms pdf spec which contains even more detail than the KQL docs page.

https://msopenspecs.azureedge.net/files/MS-KQL/%5bMS-KQL%5d.pdf

With that in mind, it looks we understood (or tried to ;)) some parts of the implicit operator node connections wrong.
There are still some parts of which contradict each other, but most of that should be now spec compliant.

following queries are now exactly covered as described in the spec:

    * 2.1.2 AND Operator 
    * 2.1.6 NOT Operator
    * 2.1.8 OR Operator
    * 2.1.12 Parentheses
    * 2.3.5 Date Tokens
        * Human tokens not implemented
    * 3.1.11 Implicit Operator
    * 3.1.12 Parentheses
    * 3.1.2 AND Operator
    * 3.1.6 NOT Operator
    * 3.1.8 OR Operator
    * 3.2.3 Implicit Operator for Property Restriction
    * 3.3.1.1.1 Implicit AND Operator
    * 3.3.5 Date Tokens

beside that it also includes:

* support for `-` and `+` bool operators
* error handling for:
    * queries that start with an binary operator
    * allow only FreeTextKeywordNodes children for GroupNodes with key
 * toolify pigeon
 * optimize grammar and parser
 * better dateTime detection/handling

## Related Issue

* https://github.com/owncloud/ocis/pull/7212
* https://github.com/owncloud/ocis/pull/7043
* https://github.com/owncloud/ocis/pull/7247
* https://github.com/owncloud/ocis/pull/7248
* https://github.com/owncloud/ocis/pull/7254
* https://github.com/owncloud/web/pull/9653
* https://github.com/owncloud/web/pull/9672
* https://github.com/owncloud/ocis/issues/7042
* https://github.com/owncloud/ocis/issues/7179
* https://github.com/owncloud/ocis/issues/7114
* https://github.com/owncloud/web/issues/9636
* https://github.com/owncloud/web/issues/9646

## Motivation and Context
with the found spec, it's more clear how some parts are connected to each other, as we said KQL will be our primary query language, therefore we should respect the spec as good as we can.

## How Has This Been Tested?

* unit tests
* acceptance tests
* e2e tests

## Screenshots (if appropriate):

## Types of changes
- [X] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Technical debt
- [ ] Tests only (no source changes)

## Checklist:
- [X] Code changes
- [X] Unit tests added
- [ ] Acceptance tests added
- [ ] Documentation ticket raised: <link>