@2403905 2023-09-08
It is not related to code. 
But related to docs and it looks counter...
```
3.3.1.1.1 Implicit AND Operator
....
The following queries match the same items:
cat +dog -fox
cat AND dog AND NOT fox
```

```
3.3.1.1.2 Implicit OR Operator
....
The following queries match the same items:
cat +dog -fox
(NOT fox) AND (dog OR (dog AND cat))
```
Why do two equal expressions become different?