@jnweiger 2023-09-08
> (NOT fox) AND (dog OR (dog AND cat))

Which is reducable. Isn't it? `(dog OR (dog AND cat))`  is logically equivalent to `(dog)`
The entire "implicit OR" thing sounds like broken-by-design to me.