@fschade 2023-09-08
@2403905, writing from my mobile, I hope it makes sense, to be exact, the docs say

cat +dog -fox
(NOT fox) AND (dog OR (dog AND cat))

are the same, which translates to:

no fox items
all dog items
all cat times but none of them without also having dog

we will never be able to translate to the exact same query from the docs, but we can creat queries that mean the same.

but i wonder too, I expect 

cat and dog and not fox reading cat +dog -fox