@fschade 2023-09-09
@2403905 i had to touch the parser again, thanks to @ScharfViktor ive found a bug in which child nodes were linked with an OR instead of an AND, i found the part in the docs, implemented it and we now can decide on a node bases how connecting edges get handled.

still, the sound strange to me, but i hope the spec is correct:

```
author:"John Smith" AND author:"Jane Smith"
```

is the same as

```
author:("John Smith" "Jane Smith")
```

but 

```
author:"John Smith" filetype:docx
```

is the same as

```
author:"John Smith" OR author:"Jane Smith"
```