@fschade 2023-09-09
also some optimizations down the road.

Before:
```
❯ go test -bench=. -benchtime=10s -count 5
goos: darwin
goarch: arm64
pkg: github.com/owncloud/ocis/v2/services/search/pkg/query/kql
BenchmarkParse-10          25560            470035 ns/op          329406 B/op       7580 allocs/op
BenchmarkParse-10          25639            465516 ns/op          329405 B/op       7580 allocs/op
BenchmarkParse-10          25536            466214 ns/op          329408 B/op       7580 allocs/op
BenchmarkParse-10          25803            472425 ns/op          329405 B/op       7580 allocs/op
BenchmarkParse-10          25431            467778 ns/op          329405 B/op       7580 allocs/op
PASS
ok      github.com/owncloud/ocis/v2/services/search/pkg/query/kql       83.663s
```

After:
```
❯ go test -bench=. -benchtime=10s -count 5
goos: darwin
goarch: arm64
pkg: github.com/owncloud/ocis/v2/services/search/pkg/query/kql
BenchmarkParse-10          48206            247459 ns/op          285539 B/op       6306 allocs/op
BenchmarkParse-10          48243            247084 ns/op          285539 B/op       6306 allocs/op
BenchmarkParse-10          48393            245621 ns/op          285539 B/op       6306 allocs/op
BenchmarkParse-10          48474            246485 ns/op          285539 B/op       6306 allocs/op
BenchmarkParse-10          48990            245175 ns/op          285540 B/op       6306 allocs/op
PASS
ok      github.com/owncloud/ocis/v2/services/search/pkg/query/kql       72.235s
```